[ADD] package config, controller with basic impl. AppRoleController, repository with basic impl., service whit basic impl

+++ src/main/resources/db/migration/V4__add_relation_between_app_user_and_app_role.sql
@@ -1,5 +1,5 @@
 alter table app_user
-    add column role_id int null;
+    add column role_id bigint null;
 
 alter table app_user
     add foreign key (role_id) references app_role (id);
